UI Integration for Random Page
set ALL Random now works

TODO:
Frontend functions
cross page specifications
Update config
Backend functions
